% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callLLM.R
\name{callLLM}
\alias{callLLM}
\title{Call Large Language Model}
\usage{
callLLM(prompt)
}
\arguments{
\item{prompt}{(string) The prompt to be passed to the language model .}
}
\value{
The response from the model
}
\description{
Call a language model via API
}
\details{
This function requires two environment variables to be set:\cr
\itemize{
\item \code{LLM_API_KEY}: The API key used for authentication.\cr
\item \code{LLM_ENDPOINT}: The full URL endpoint for the API.\cr
}

Set these environment variables either in your R session using \code{Sys.setenv()}, or permanently in a \code{.Renviron} file.\cr

If either \code{LLM_API_KEY} or \code{LLM_ENDPOINT} is not set, the function will throw an error.\cr
}
