% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEligibilityLogic.R
\name{checkEligibilityLogic}
\alias{checkEligibilityLogic}
\title{Check eligibility logic generated by LLM}
\usage{
checkEligibilityLogic(eligibility_data)
}
\arguments{
\item{eligibility_data}{A dataframe including a logic column that expresses trial's eligibility as an R expression}
}
\value{
A dataframe with the following additional columns:
\itemize{
\item \code{valid_logic} - (\code{TRUE} or \code{FALSE}) - does the expression evaluate a test patient without error?
\item \code{whitelist} - (\code{TRUE} or \code{FALSE}) - does the expression consist entirely of valid tokens?
}
}
\description{
\itemize{
\item Check whether the R expression generated by LLM to determine patient eligibility evaluates an example patient object without throwing an error
\item Because of the potential security risk of running R code that has been generated by LLM, check the content of the expression.
}
}
\details{
The only allowed tokens are:
\itemize{
\item \code{!} (logical NOT)
\item \code{\%in\%} (logical NOT)
\item \code{&} (logical AND)
\item \code{|} (logical OR)
\item \code{any} (logical NOT)
\item \code{patient$condition} (logical NOT)
\item \code{patient$alteration} (logical NOT)
\item \code{C} followed by one or more digits
}
}
\seealso{
\code{\link{createEligibilityLogic}}
}
