% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPrompt1Outputs.R
\name{checkPrompt1Outputs}
\alias{checkPrompt1Outputs}
\title{Check whether first prompts have already been used}
\usage{
checkPrompt1Outputs(latest_prompts, db_path)
}
\arguments{
\item{latest_prompts}{A dataframe that includes prompt1 values formed during the previous step.}

\item{db_path}{Path to a SQLite database - if the database includes an \code{eligibility} table, this function will check if a prompt has been used before}
}
\value{
A dataframe with an additional column \code{prompt_1_output}. This column is filled with NA by default, but includes previous responses if prompt1 has already been sent previously.
}
\description{
If the pipeline has been run before, check for each trial whether an identical prompt1 has been sent to LLM before.\cr
If an identical prompt has been sent previously, use the output from that call in order to reduce the number of calls that will go the the LLM during the next step of the pipeline.\cr
}
