% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_trial_json.R
\name{compose_trial_json}
\alias{compose_trial_json}
\title{Convert trial data to json format and lookup cancer types and alterations using dictionaries}
\usage{
compose_trial_json(trial_data)
}
\arguments{
\item{trial_data}{A list of dataframes containing raw clinical trial data.\cr}
}
\value{
A dataframe, 1 row per trial including JSON-structured data\cr
}
\description{
As a first step towards annotating trial eligibility, convert relevant data to json format, 1 json string per trial.\cr
Relevant data include condition names, design group titles and descriptions, and eligibility criteria.\cr
}
