% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEligibilityLogic.R
\name{createEligibilityLogic}
\alias{createEligibilityLogic}
\title{Use LLM to generate R expressions that evaluate trial eligibility}
\usage{
createEligibilityLogic(eligibility_data)
}
\arguments{
\item{eligibility_data}{A dataframe including columns \code{prompt_2} and \code{prompt_2_output}}
}
\value{
The input dataframe with the \code{prompt_2_output} column updated with the generated logic from the LLM. For treatment arms that already have an output (or where no relevant terms are mentioned (because an identical prompt2 has been used before) the original response is preserved.
}
\description{
Based on the prompts formed in the previous step, generate R expressions (1 per treatment arm) that evaluate a patient object and return either \code{TRUE} (eligibile) or \code{FALSE} (ineligible)
}
\details{
Calls the function `callLLM' for which API endpoint and key are required. These are specified as environment variables
}
\seealso{
\code{\link{callLLM}}
}
