% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_LLM_lookup.R
\name{eligibility_LLM_lookup}
\alias{eligibility_LLM_lookup}
\title{Additional lookups of cancer types and alterations using LLM}
\usage{
eligibility_LLM_lookup(eligibility_data, dictionaries)
}
\arguments{
\item{eligibility_data}{A dataframe containing trial information, including columns \code{prompt_1} and \code{prompt_1_output} for storing the LLM-processed output.}

\item{dictionaries}{A list containing two specific dictionaries: \code{cancer_dict} for cancer types and \code{alteration_dict} for molecular dysfunctions.}
}
\value{
Returns the input dataframe with additional columns:
\itemize{
\item \code{cancer_lookup_llm}: for each trial, a list of NCI thesaurus codes related to cancer types.
\item \code{moldysf_lookup_llm}: for each trial, a list of NCI thesaurus codes related to molecular dysfunctions.
}
}
\description{
Prompt language model with trial data to get cancer types and alterations mentioned in trial record.\cr
This is just a lookup, there is no determination of how these mentions relate to eligibility.\cr
For each trial, if an identical prompt has already been sent previously and the results stored in a local database, the result from that call will be used, and the language model will not be called again.
The outputs from the LLM (whether obtained previously or through this function) will be used as the basis for dictionary lookups in order to map to NCI thesaurus codes.
}
\details{
Calls the function `callLLM' for which API endpoint and key are required. These are specified as environment variables
}
\seealso{
\code{\link{callLLM}}
}
