% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibilityLookup.R
\name{eligibilityLookup}
\alias{eligibilityLookup}
\title{Convert trial data to json format and lookup cancer types and alterations using dictionaries}
\usage{
eligibilityLookup(json_data, dictionaries)
}
\arguments{
\item{json_data}{A dataframe that includes trial data in column \code{json}, in json format.\cr}

\item{dictionaries}{A list of dictionaries including \code{cancer_dict} and \code{alteration_dict}.\cr}
}
\value{
A dataframe, 1 row per trial including new columns with results of lookups:\cr
\itemize{
\item \code{cancer_lookup}
\item \code{moldysf_lookup}
}
}
\description{
As a first step towards annotating trial eligibility, convert relevant data to json format, 1 json string per trial.\cr
Relevant data include condition names, design group titles and descriptions, and eligibility criteria.\cr
Lookup matches against dictionaries of cancer types and molecular alterations in the json strings.\cr
}
