% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formPrompt1.R
\name{formPrompt1}
\alias{formPrompt1}
\title{Form first prompts to go to Language Model}
\usage{
formPrompt1(eligibility_data)
}
\arguments{
\item{eligibility_data}{A dataframe containing trial information in JSON format.}
}
\value{
The input dataframe with an additional column, \code{prompt_1}, which contains the fully formed prompts for the LLM.
}
\description{
In addition to dictionary based lookups of cancer types and alterations in trial data, form prompts (1 per trial) that will send the trial data to language model in order to identify additional matches.\cr
}
\details{
\itemize{
\item Uses few-shot prompting with chain of thought reasoning.\cr
\item Prompts are formed by concatenating instructions, examples, and trial data.\cr
\item Instructions and examples are stored at the following relative paths:\cr
\itemize{
\item \code{inst/prompts/prompt_1_instructions.txt}
\item \code{dinst/prompts/prompt_1_example_1.txt}
\item \code{inst/prompts/prompt_1_example_2.txt}
\item \code{inst/prompts/prompt_1_example_3.txt}
}
}
}
