% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formPrompt2.R
\name{formPrompt2}
\alias{formPrompt2}
\title{Form second prompts to go to language model}
\usage{
formPrompt2(eligibility_data, thesaurus)
}
\arguments{
\item{eligibility_data}{A dataframe containing the initial eligibility criteria data
and the results from both LLM processing and dictionary lookups for cancer types and
molecular dysfunctions.}

\item{thesaurus}{A list containing the pruned thesaurus data used for obtaining the
Preferred Terms (PTs) corresponding to the identified codes from the lookups.}
}
\value{
Returns the input dataframe augmented with additional columns:
\itemize{
\item \code{cancer_lookups_combined}: A combined list of unique cancer type codes identified from both LLM outputs and dictionary lookups.
\item \code{moldysf_lookups_combined}: A combined list of unique molecular dysfunction codes identified from both LLM outputs and dictionary lookups.
\item \code{cancer_codes_json}: JSON-formatted strings containing the PTs for the combined cancer type codes.
\item \code{moldysf_codes_json}: JSON-formatted strings containing the PTs for the combined molecular dysfunction codes.
\item \code{prompt_2}: The fully formed second prompts, incorporating instructions, examples, and the trial information to be analyzed, along with the potential eligible conditions and alterations in JSON format.
}
}
\description{
For each trial, form a second prompt that will be sent to the language model.\cr
The second prompt will instruct the language model to generate an R expression (1 per treatment arm) that evaluates a patient object and returns TRUE (i.e. eligible) or FALSE (i.e. ineligible)\cr
Each prompt will include:\cr
\itemize{
\item Instructions\cr
\item Examples (with chain of thought reasoning)\cr
\item A set of cancer types (and their NCI thesaurus codes) that are potentially relevant to eligibility\cr
\item A set of alterations (and their NCI thesaurus codes) that are potentially relevant to eligibilty\cr
\item The trial information (json format)\cr
}
}
\details{
\itemize{
\item Uses few-shot prompting with chain of thought reasoning.\cr
\item Prompts are formed by concatenating instructions, examples, and trial data.\cr
\item Instructions and examples are configurable, and should be stored as text files in the following relative paths:\cr
\itemize{
\item \code{inst/prompts/prompt_2_instructions.txt}
\item \code{inst/prompts/prompt_2_example_1.txt}
\item \code{inst/prompts/prompt_2_example_2.txt}
}
}
}
