% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnestEligibilityLogic.R
\name{unnestEligibilityLogic}
\alias{unnestEligibilityLogic}
\title{Unnest and Transform Eligibility Logic from JSON to Data Frame Format}
\usage{
unnestEligibilityLogic(eligibility_data, raw_trial_data)
}
\arguments{
\item{eligibility_data}{A dataframe including \code{prompt_2_output} column storing JSON-formatted eligibility logic for each trial.}

\item{raw_trial_data}{A dataframe containing raw trial data, including design groups information to be linked to the eligibility logic.}
}
\value{
A dataframe similar to the input \code{eligibility_data} but with the JSON-formatted \code{prompt_2_output} converted and unnested into separate columns.
Additional columns for design group titles (\code{dg_title}), statuses (\code{dg_status}), and IDs (\code{dg_id}) are also included
}
\description{
Converts the LLM output from prompt2 from json format into dataframe and unnests (empty rows are retained).
}
\details{
Note that design group IDs are not static - i.e. the same design group could have different IDs over time - this is a feature of the AACT database.
}
